refactor: apply remaining clippy suggestions for unused variables
Fix additional clippy warnings:
- Remove unused variable warnings by prefixing with underscore
- Simplify never-loop pattern in seedfill
- Adjust match arm formatting in watershed and scale
- Clean up unnecessary variable bindings

+++ crates/leptonica-color/src/analysis.rs
@@ -336,7 +336,7 @@ mod tests {
 
         for y in 0..10 {
             for x in 0..10 {
-                let val = (x * 25) as u32;
+                let val = x * 25;
                 unsafe { pix_mut.set_pixel_unchecked(x, y, val) };
             }
         }
